server/requirements.txt
fastmcp 
ollama 
requests
psycopg2-binary
flask
loguru
python-dotenv
duckdb
pandas
groq
mysql-connector-python